Add one more type of color parsing (string only)
