--- ClosePath.cpp
@@ -1,6 +1,0 @@
-#include "ClosePath.h"
-using namespace Gdiplus;
-
-void ClosePath::draw(Graphics* graphics, GraphicsPath* path) {
-    path->CloseFigure();
-}
