--- LineTo.cpp
@@ -1,8 +1,0 @@
-#include "LineTo.h"
-using namespace Gdiplus;
-
-void Lineto::draw(Graphics* graphics, GraphicsPath* path) {
-    PointF last;
-    path->GetLastPoint(&last);
-    path->AddLine(last, PointF(destination.X, destination.Y));
-}
