--- Path.h
@@ -1,39 +1,0 @@
-#ifndef _SVGPATH_
-#define _SVGPATH_
-
-#include <objidl.h>
-#include <windows.h>
-#include <gdiplus.h>
-#include <sstream>
-#include <vector>
-#include <string>
-#include <sstream>
-#include "tinyxml2.h"
-
-
-
-#include "SVGElement.h"
-#include "MoveTo.h"
-#include "CurveTo.h"
-#include "LineTo.h"
-#include "HLineTo.h"
-#include "VLineTo.h"
-#include "ClosePath.h"
-
-
-using namespace std;
-using namespace Gdiplus;
-
-#pragma comment (lib,"Gdiplus.lib")
-
-class path : public SVGElement {
-    vector<PathCommand*> commands;
-public:
-    path();
-    ~path();
-    void setValue(tinyxml2::XMLElement* element) override;
-    void draw(HDC hdc) override;
-    void draw(Graphics* g) override;
-};
-
-#endif // !_SVGPATH_
