+++ SVG_Project/ClosePath.cpp
@@ -1,0 +1,6 @@
+#include "ClosePath.h"
+using namespace Gdiplus;
+
+void ClosePath::draw(Graphics* graphics, GraphicsPath* path) {
+    path->CloseFigure();
+}
