+++ SVG_Project/ClosePath.h
@@ -1,0 +1,11 @@
+#ifndef _CLOSEPATH_
+#define _CLOSEPATH_
+
+#include "PathCommand.h"
+
+class ClosePath : public PathCommand {
+public:
+    void draw(Gdiplus::Graphics* graphics, Gdiplus::GraphicsPath* path) override;
+};
+
+#endif
