+++ SVG_Project/LineTo.cpp
@@ -1,0 +1,8 @@
+#include "LineTo.h"
+using namespace Gdiplus;
+
+void Lineto::draw(Graphics* graphics, GraphicsPath* path) {
+    PointF last;
+    path->GetLastPoint(&last);
+    path->AddLine(last, PointF(destination.X, destination.Y));
+}
