--- VLineTo.h
@@ -1,13 +1,0 @@
-#ifndef _VLINETO_
-#define _VLINETO_
-
-#include "PathCommand.h"
-
-class VLineTo : public PathCommand {
-    int y;
-public:
-    VLineTo(int yCoord) : y(yCoord) {}
-    void draw(Gdiplus::Graphics* graphics, Gdiplus::GraphicsPath* path) override;
-};
-
-#endif
